// frontend/src/api/generated/model/tutor-request-response-dto.ts
/**
 * EduMatch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TutorRequestResponseDto { 
    id?: number;
    code?: string | null;
    studentId?: number;
    studentName?: string | null;
    studentAvatar?: string | null;
    subjectId?: number;
    subjectName?: string | null;
    note?: string | null;
    status?: string | null;
    isExpired?: boolean;
    pricePerSession?: number;
    fullAddress?: string | null;
    expiresAt?: string | null;
    preferredSchedule?: string | null;
    sessionsPerWeek?: number;
    minutesPerSession?: number;
    sessionsPerMonth?: number;
    gradeLevel?: string | null;
    educationLevel?: string | null;
    applicationCount?: number;
    createdAt?: string;
}